requirements.txt
scrapling
sqlite3
argparse
logging
datetime
json
re
python-dotenv
